refactor(characterArchetypes): extract static template data to constants
Extracted static archetype templates and physical description data from characterArchetypes.ts to a dedicated constants file. This separates data from logic for better maintainability.

Changes:
- Created src/lib/constants/characterArchetypeTemplates.ts (472 lines)
- Reduced characterArchetypes.ts from 1,026 lines to 330 lines (698 line reduction, 68%)
- Replaced large switch statement with simple lookup pattern
- Maintained all generator logic and type safety
- Updated tests to reflect fantasy fallback for unknown genres
- All 1,815 tests passing

Related to issue #713 (Phase 2: characterArchetypes consolidation)
